Refactored ThumbsView with shared base class.
