+++ src/modules/uv-shared-module/css/thumbs-view.less
@@ -1,0 +1,132 @@
+#app{
+
+    .leftPanel{
+
+        // TODO: move/copy images from contentleftpanel to shared
+        @img-path: '../../../modules/uv-contentleftpanel-module/img/';
+
+        .views{
+
+            .thumbsView{
+
+                overflow: auto;
+
+                .thumbs {
+                    overflow: hidden;
+                    width: 210px;
+                    position: relative;
+
+                    .thumb {
+                        cursor: pointer;
+
+                        .wrap {
+                            border: 2px solid @panel-border-color;
+                            padding: 2px;
+                            overflow: hidden;
+                            text-align: center;
+                            background: none;
+
+                            &.loading {
+                                .loading(@loader-black-bg);
+                            }
+
+                            &.loadingFailed {
+                                background: url('@{img-path}error.png') no-repeat 50% 50%;
+                            }
+
+                            &.hidden {
+                                background: url('@{img-path}hidden_thumb.png') no-repeat 50% 50%;
+                            }
+
+                            img {
+                                display: block;
+                                max-width: 100%;
+                                max-height: 100%;
+                                //margin: 0 auto;
+                            }
+                        }
+
+                        &.selected {
+
+                            .wrap {
+                                border: 2px solid #fff;
+                            }
+                        }
+
+                        &.searchpreview{
+
+                            .wrap{
+                                border: 2px solid @link-color;
+                            }
+                        }
+
+                        span {
+                            font-size: @font-size-small;
+                            color: #fff;
+                            display: block;
+                            padding: 2px 0 0 0;
+                        }
+
+                        .label {
+                            overflow-x: hidden;
+                            text-overflow: ellipsis;
+                            white-space: nowrap;
+                        }
+
+                        &.placeholder {
+                            .wrap {
+                                background: url('@{img-path}unavailable.png') no-repeat 50% 50%;
+                            }
+                        }
+
+                        &.oneCol{
+                            margin: 0 0 7px 0;
+                            .wrap,
+                            .label {
+                                width: 208px;
+                                max-height: 373px;
+                            }
+                        }
+
+                        &.twoCol{
+                            margin: 0 7px 7px 0;
+                            .wrap,
+                            .label {
+                                width: 98px;
+                                max-height: 182px;
+                            }
+                        }
+                    }
+
+                    .separator {
+                        height: 1px;
+                        clear: both;
+                        border: none;
+                    }
+
+                    &.left-to-right{
+                        &.paged{
+                            .thumb{
+                                &.first{
+                                    float: right;
+                                }
+                            }
+                        }
+                        .thumb{
+                            float: left;
+                        }
+                    }
+
+                    &.right-to-left{
+                        .thumb{
+                            &.first{
+                                float: left;
+                            }
+                            float: right;
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
